+++ cups/cups/ppdc/ppdc.h
@@ -1,14 +1,10 @@
 //
 // Definitions for the CUPS PPD Compiler.
 //
-// Copyright 2007-2009 by Apple Inc.
+// Copyright 2007-2019 by Apple Inc.
 // Copyright 2002-2007 by Easy Software Products.
 //
-// These coded instructions, statements, and computer programs are the
-// property of Apple Inc. and are protected by Federal copyright
-// law.  Distribution and use rights are outlined in the file "LICENSE.txt"
-// which should have been included with this file.  If this file is
-// missing or damaged, see the license at "http://www.cups.org/".
+// Licensed under Apache License v2.0.  See the file "LICENSE" for more information.
 //
 
 #ifndef _PPDC_H_
@@ -109,7 +105,7 @@ class ppdcArray				//// Shared Array
 {
   public:
 
-  int		count,			// Number of elements
+  size_t	count,			// Number of elements
 		alloc,			// Allocated elements
 		current;		// Current element
   ppdcShared	**data;			// Elements
