+++ ppp/Family/ppp_link.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2000 Apple Computer, Inc. All rights reserved.
+ * Copyright (c) 2000, 2018 Apple Inc. All rights reserved.
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
@@ -448,14 +448,14 @@ void ppp_link_logmbuf(struct ppp_link *link, char *msg, mbuf_t m)
     IOLog("%s: [ifnet = %s%d] [link = %s%d]\n", msg,
             LKIFNAME(link), LKIFUNIT(link), LKNAME(link), LKUNIT(link));
 
-    for (count = mbuf_len(m), data = mbuf_data(m); m != NULL; ) {	// no alignment issue as data is *uchar.
+    for (count = (int)mbuf_len(m), data = mbuf_data(m); m != NULL; ) {	// no alignment issue as data is *uchar.
         /* build a line of output */
         for(lcount = 0; lcount < sizeof(lbuf); lcount += copycount) {
             if (!count) {
                 m = mbuf_next(m);
                 if (m == NULL)
                     break;
-                count = mbuf_len(m);
+                count = (int)mbuf_len(m);
                 data  = mbuf_data(m);
             }
             copycount = (count > sizeof(lbuf) - lcount) ? sizeof(lbuf) - lcount : count;
